................
r43898 | kpfleming | 2006-09-28 11:38:25 -0500 (Thu, 28 Sep 2006) | 10 lines

Merged revisions 43895 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r43895 | kpfleming | 2006-09-28 11:32:44 -0500 (Thu, 28 Sep 2006) | 2 lines

eliminate compiler warning introduced by recent changes

........

................


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@43900 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/cli.c
@@ -743,7 +743,7 @@ static int handle_showchan(int fd, int argc, char *argv[])
 		c->fout & ~DEBUGCHAN_FLAG, (c->fout & DEBUGCHAN_FLAG) ? " (DEBUGGED)" : "",
 		(long)c->whentohangup,
 		cdrtime, c->_bridge ? c->_bridge->name : "<none>", ast_bridged_channel(c) ? ast_bridged_channel(c)->name : "<none>", 
-		c->context, c->exten, c->priority, c->callgroup, c->pickupgroup, ( c->appl ? c->appl : "(N/A)" ),
+		c->context, c->exten, c->priority, (int) c->callgroup, (int) c->pickupgroup, ( c->appl ? c->appl : "(N/A)" ),
 		( c-> data ? S_OR(c->data, "(Empty)") : "(None)"),
 		(ast_test_flag(c, AST_FLAG_BLOCKING) ? c->blockproc : "(Not Blocking)"));
